+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
     name: 'pwix:modal',
-    version: '1.8.1-rc',
+    version: '1.9.0-rc',
     summary: 'A Bootstrap-based package which provides draggable and resizable modal dialogs to Meteor',
     git: 'https://github.com/trychlos/pwix-modal',
     documentation: 'README.md'
@@ -31,6 +31,7 @@ function configure( api ){
     api.use( 'pwix:jquery-ui@1.0.2' );
     api.use( 'pwix:layout@1.3.0' );
     api.use( 'random', 'client' );
+    api.use( 'reactive-var', 'client' );
     api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
 }