+++ src/client/classes/md_button.class.js
@@ -20,7 +20,7 @@ export class mdButton {
      * @summary Define the initial set of buttons of a modal
      *  set 'last=true' on the last button
      * @param {String|Object|Array} defs the definition of the buttons
-     * @returns {Array} the list of defined buttons
+     * @returns {Array} the list of defined buttons, each one as an independant reactive var which contains a mdButton
      */
     static define( defs ){
         //console.debug( 'define input', defs );
@@ -42,7 +42,8 @@ export class mdButton {
         });
         // flag the last object (will be displayed on rightest side)
         if( buttons.length ){
-            buttons[buttons.length-1].last = true;
+            const btn = buttons[buttons.length-1];
+            btn.last = true;
         }
         return buttons;
     }
@@ -71,62 +72,89 @@ export class mdButton {
 
     /**
      * @summary Update one or several existing button(s)
-     * @param {Array} defined the current definition of the buttons
-     * @param {String|Object|Array} def the definition of a button
+     * @param {mdModal} modal the calling mdModal
+     * @param {String|Object|Array} defs the definition of one or more buttons
+     *  Contrarily to the initial definition, this update MUST have an id in each provided object
      * @returns {Array} the updated defined list
      */
-    static update( defined, sets ){
-        //console.debug( 'defined', defined );
-        //console.debug( 'sets', sets );
-        let asets = _.isArray( sets ) ? sets : [ sets ];
-        asets.every(( set ) => {
-            //console.debug( 'set', set );
-            if( !set.id ){
+    static update( modal, defs ){
+        let adefs = _.isArray( defs ) ? defs : [ defs ];
+        adefs.every(( def ) => {
+            if( !def.id ){
                 console.error( 'identifier is mandatory, not found' );
+            } else if( def.id === Modal.C.ButtonExt.RESET ){
+                modal.buttonsReset();
             } else {
-                let found = null;
-                defined.every(( def ) => {
-                    //console.debug( 'def', def );
-                    if( def.id === set.id ){
-                        found = def;
-                    }
-                    return found === null;
-                });
-                if( !found ){
-                    const btn = mdButton.new( set );
-                    if( btn ){
-                        if( defined.length ){
-                            defined[defined.length-1].last = false;
-                        }
-                        btn.last = true;
-                        defined.push( btn );
-                    }
+                let found = modal.buttonGet( def.id );
+                if( found ){
+                    found._setDef( def );
                 } else {
-                    _.merge( found, set );
-                    // setup the reactive vars
-                    if( Object.keys( set ).includes( 'enabled' ) && _.isBoolean( set.enabled )){
-                        found._enabled.set( set.enabled );
+                    const btn = mdButton.new( def );
+                    if( btn ){
+                        modal.buttonAdd( btn );
                     }
                 }
             }
             return true;
         });
-        return defined;
     }
 
     // private data
     //
 
+    _cb = new ReactiveVar( null );
+    _classes = new ReactiveVar( null );
+    _dismiss = new ReactiveVar( null );
     _enabled = new ReactiveVar( true );
+    _html = new ReactiveVar( null );
+    _label = new ReactiveVar( null );
+    _name = new ReactiveVar( null );
+    _type = new ReactiveVar( 'button' );
 
     // private methods
     //
+
+    _setBool( arg, def ){
+        if( arg === true || arg == false ){
+            return arg;
+        }
+        if( def === true || def === false ){
+            return def;
+        }
+        return undefined;
+    }
+
+    _setDef( def ){
+        if( Object.keys( def ).includes( 'cb' )){
+            this._cb.set( def.cb );
+        }
+        if( Object.keys( def ).includes( 'classes' )){
+            this._classes.set( def.classes );
+        }
+        if( Object.keys( def ).includes( 'dismiss' ) && ( def.dismiss === true || def.dismiss === false )){
+            this._dismiss.set( def.dismiss );
+        }
+        if( Object.keys( def ).includes( 'enabled' ) && ( def.enabled === true || def.enabled === false )){
+            this._enabled.set( def.enabled );
+        }
+        if( Object.keys( def ).includes( 'html' )){
+            this._html.set( def.html );
+        }
+        if( Object.keys( def ).includes( 'label' )){
+            this._label.set( def.label );
+        }
+        if( Object.keys( def ).includes( 'name' )){
+            this._name.set( def.name );
+        }
+        if( Object.keys( def ).includes( 'type' )){
+            this._type.set( def.type );
+        }
+    }
 
     // public data
     //
 
     id = null;
-    parms = null;
 
     /**
      * Constructor
@@ -135,28 +163,36 @@ export class mdButton {
      */
     constructor( def ){
         this.id = def.id;
-        this.parms = def;
 
-        // initialize the enabled rv
-        if( Object.keys( def ).includes( 'enabled' ) && _.isBoolean( def.enabled )){
-            this._enabled.set( def.enabled );
+        this._setDef( def );
+        if( !this._name.get()){
+            this._name.set( this.id );
         }
 
+        //console.debug( this );
         return this;
+    }
+
+    /**
+     * @return {Function} the calback function associated to the button
+     */
+    cb(){
+        return this._cb.get();
     }
 
     /**
      * @return {String} the classes to set to the button
      */
     classes(){
-        return this.parms.classes || ( this.last ? 'btn-primary' : 'btn-secondary' );
+        return this._classes.get() || ( this.last ? 'btn-primary' : 'btn-secondary' );
     }
 
     /**
      * @return {Boolean} whether the button dismiss the dialog
      */
     dismiss(){
-        return Object.keys( this.parms ).includes( 'dismiss' ) && _.isBoolean( this.parms.dismiss ) ? this.parms.dismiss : undefined;
+        const dismiss = this._dismiss.get();
+        return dismiss === true || dismiss === false ? dismiss : undefined;
     }
 
     /**
@@ -170,27 +206,27 @@ export class mdButton {
      * @return {String} the HTML definition of the button, or ''
      */
     html(){
-        return this.parms.html || '';
+        return this._html.get() || '';
     }
 
     /**
      * @return {String} the label of the button
      */
     label(){
-        return this.parms.label || ( Object.keys( Modal.C.Button ).includes( this.id ) ? pwixI18n.label( I18N, this.id ) : this.id );
+        return this._label.get() || ( Object.keys( Modal.C.Button ).includes( this.id ) ? pwixI18n.label( I18N, this.id ) : this.id );
     }
 
     /**
      * @return {String} the name of the button
      */
     name(){
-        return this.parms.name || this.id;
+        return this._name.get() || this.id;
     }
 
     /**
      * @return {String} the type of the button
      */
     type(){
-        return this.parms.type || 'button';
+        return this._type.get() || 'button';
     }
 }